add sex validation to rna fusion
