+++ cg/services/order_validation_service/rules/case_sample/rules.py
@@ -21,6 +21,7 @@ from cg.services.order_validation_service.errors.case_sample_errors import (
     PedigreeError,
     SampleDoesNotExistError,
     SampleNameRepeatedError,
+    SexSubjectIdError,
     SubjectIdSameAsCaseNameError,
     SubjectIdSameAsSampleNameError,
     VolumeRequiredError,
@@ -41,6 +42,7 @@ from cg.services.order_validation_service.rules.case_sample.utils import (
     get_mother_sex_errors,
     get_occupied_well_errors,
     get_well_sample_map,
+    has_sex_and_subject,
     is_concentration_missing,
     is_container_name_missing,
     is_invalid_plate_well_format,
@@ -324,6 +326,30 @@ def validate_subject_ids_different_from_case_names(
             case_index=index,
         )
         errors.extend(case_errors)
+    return errors
+
+
+def validate_subject_sex_consistency(
+    order: OrderWithCases,
+    store: Store,
+) -> list[SexSubjectIdError]:
+    errors: list[SexSubjectIdError] = []
+
+    for case_index, case in order.enumerated_new_cases:
+        for sample_index, sample in case.enumerated_new_samples:
+            if not has_sex_and_subject(sample):
+                continue
+
+            if store.sample_exists_with_different_sex(
+                customer_internal_id=order.customer,
+                subject_id=sample.subject_id,
+                sex=sample.sex,
+            ):
+                error = SexSubjectIdError(
+                    case_index=case_index,
+                    sample_index=sample_index,
+                )
+                errors.append(error)
     return errors
