+++ cg/services/order_validation_service/workflows/balsamic/rules/case_sample/utils.py
@@ -1,14 +1,7 @@
 from cg.constants import PrepCategory
-from cg.models.orders.sample_base import SexEnum
-from cg.services.order_validation_service.workflows.balsamic.models.sample import (
-    BalsamicSample,
-)
+from cg.services.order_validation_service.workflows.balsamic.models.sample import BalsamicSample
 from cg.store.models import Application
 from cg.store.store import Store
-
-
-def has_sex_and_subject(sample: BalsamicSample) -> bool:
-    return bool(sample.subject_id and sample.sex != SexEnum.unknown)
 
 
 def is_sample_missing_capture_kit(sample: BalsamicSample, store: Store) -> bool:
